Add rotate and resize for image feature

+++ src/main/java/com/jalasoft/jfc/model/strategy/CommandImageRotate.java
@@ -21,13 +21,13 @@ import com.jalasoft.jfc.model.pdf.ImageMagickCommand;
 public class CommandImageRotate implements ICommandStrategy {
 
     // Content command value.
-    private int commandValue;
+    private double commandValue;
 
     /**
      * It creates a new CommandImageRotate object.
      * @param commandValue contains a value.
      */
-    public CommandImageRotate(int commandValue) {
+    public CommandImageRotate(double commandValue) {
         this.commandValue = commandValue;
     }
